chore: rpc -> api-model plumbing for nvl_logical_partition
This continues the work from https://github.com/NVIDIA/ncx-infra-controller-core/pull/606, https://github.com/NVIDIA/ncx-infra-controller-core/pull/602, https://github.com/NVIDIA/ncx-infra-controller-core/pull/598, https://github.com/NVIDIA/ncx-infra-controller-core/pull/596, and https://github.com/NVIDIA/ncx-infra-controller-core/pull/608.

TLDR is we've leaked a few things from `::rpc` into the `api-db` layer, which we generally don't want to do, and now that we have `STYLE_GUIDE.md`, it will be good to practice what we preach.

This specific change introduces `nvl_logical_partition` equivalents in `api-model`, with `From`/`TryFrom` impls for the API layer to translate to before passing downstream.
