+++ crates/api/src/handlers/logical_partition.rs
@@ -16,9 +16,9 @@
  */
 use ::rpc::forge as rpc;
 use config_version::ConfigVersion;
-use db::nvl_logical_partition::{self, NewLogicalPartition};
-use db::{self, ObjectColumnFilter, WithTransaction, instance};
+use db::{self, ObjectColumnFilter, WithTransaction, instance, nvl_logical_partition};
 use futures_util::FutureExt;
+use model::nvl_logical_partition::NewLogicalPartition;
 use tonic::{Request, Response, Status};
 
 use crate::CarbideError;
@@ -44,7 +44,9 @@ pub(crate) async fn create(
     let metadata = req.config.metadata.clone();
     metadata.validate(true).map_err(CarbideError::from)?;
 
-    let resp = req.create(&mut txn).await.map_err(CarbideError::from)?;
+    let resp = nvl_logical_partition::create(&req, &mut txn)
+        .await
+        .map_err(CarbideError::from)?;
     let resp = rpc::NvLinkLogicalPartition::try_from(resp).map(Response::new)?;
     txn.commit().await?;
